pipeline/requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.2.2
joblib==1.4.2
boto3==1.34.101
botocore==1.34.101
PyYAML==5.1.1